+++ app/src/main/java/com/example/mylifearranger/feature_planner/data/repository/PlanRepositoryImpl.kt
@@ -15,7 +15,7 @@ class PlanRepositoryImpl(
         return planDao.getPlanWithTasks(planId)
     }
 
-    override suspend fun insertPlan(plan: Plan): Int {
+    override suspend fun insertPlan(plan: Plan): Long {
         return planDao.insertPlan(plan)
     }
